Report progress notifications every 5% increments
Fixes https://github.com/scalacenter/bloop/issues/788

+++ backend/src/main/scala/bloop/Compiler.scala
@@ -63,7 +63,7 @@ object Compiler {
     override def advance(current: Int, total: Int): Boolean = {
       val isNotCancelled = !cancelPromise.isCompleted
       if (isNotCancelled) {
-        reporter.reportCompilationProgress(current, total, lastPhase, lastUnitPath)
+        reporter.reportCompilationProgress(current.toLong, total.toLong, lastPhase, lastUnitPath)
       }
 
       isNotCancelled
